rename DataLayerClient/DataLayerService to DataLayerClientService/DataLayerServerService
